bug fix
git-svn-id: http://192.168.110.32/svn/Products/PFF/Base/Branches/Release-2.41.0@73047 6963f9ea-81fa-44da-aa10-7ebde50aa541

+++ PFSJava/src/main/java/com/pennant/backend/service/finance/impl/FinanceMaintenanceServiceImpl.java
@@ -816,9 +816,11 @@ public class FinanceMaintenanceServiceImpl extends GenericFinanceDetailService i
 
 		FinODPenaltyRate oldPenaltyRate = getFinODPenaltyRateDAO().getFinODPenaltyRateByRef(
 				financeMain.getFinReference(), "");
-		getFinODPenaltyRateDAO().saveLog(oldPenaltyRate, "_Log");
-		getFinODPenaltyRateDAO().delete(financeMain.getFinReference(), "");
-		getFinODPenaltyRateDAO().save(penaltyRate, "");
+		if(oldPenaltyRate!=null){
+			getFinODPenaltyRateDAO().saveLog(oldPenaltyRate, "_Log");
+			getFinODPenaltyRateDAO().delete(financeMain.getFinReference(), "");
+			getFinODPenaltyRateDAO().save(penaltyRate, "");
+		}
 
 		// Save Document Details
 		if (financeDetail.getDocumentDetailsList() != null && financeDetail.getDocumentDetailsList().size() > 0) {
